refactor: centralize odds conversion
